♻️ TSE-05-enums : Migrate js enums to ts + clarify usage rename enums to constants
