+++ src/components/pages/order/Main/MainRightSide/Admin/AdminPanel/AdminPanel.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { useOrderContext } from "../../../../../../../context/OrderContext";
-import { EMPTY_PRODUCT } from "../../../../../../../enums/product";
+import { EMPTY_PRODUCT } from "../../../../../../../constants/product";
 import { theme } from "../../../../../../../theme/themeSystem";
 import { getTabSelected, getTabsConfig } from "../tabsConfig";
