+++ src/components/pages/order/Main/MainRightSide/Menu/Menu.jsx
@@ -10,7 +10,7 @@ import {
   EMPTY_PRODUCT,
   IMAGE_COMING_SOON,
   IMAGE_NO_STOCK,
-} from "../../../../../../enums/product";
+} from "../../../../../../constants/product";
 import { isEmpty } from "../../../../../../utils/array";
 import Loader from "./Loader";
 import { CSSTransition, TransitionGroup } from "react-transition-group";
