+++ src/constants/product.ts
@@ -1,11 +1,13 @@
-export const EMPTY_PRODUCT = Object.freeze({
+import { MenuProduct } from "@/types/Product";
+
+export const EMPTY_PRODUCT: MenuProduct = {
   id: "",
-  title: "",
+  title: "Produit vide",
   imageSource: "",
-  price: 0,
+  price: "0",
   isAvailable: true,
   isPublicised: false,
-});
+} as const;
 
 export const IMAGE_COMING_SOON: string = "/images/coming-soon.png";
 export const IMAGE_NO_STOCK: string = "/images/stock-epuise.png";
@@ -14,4 +16,4 @@ export const BASKET_MESSAGE = {
   EMPTY: "Votre commande est vide.",
   LOADING: "Chargement en cours...",
   NOT_AVAILABLE: "Non disponible",
-};
+} as const;
